requirements.txt
requests>=2.31.0
fpdf2>=2.7.9
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-chroma>=0.2.0
langchain-huggingface>=0.1.0
sentence-transformers>=3.0.0
langchain-text-splitters>=0.3.0
chromadb>=0.5.0
pypdf>=4.0.0
pytest>=8.0.0
